Prepare for new implementation of computing inverse
